Update maintain-guides-how-to-validate-kusama.md
Noticed that I was failing on multiple systems because I did not have the binary for `protoc`. 
Note that I am installing on a Raspberry Pi 4 with Raspbian OS 32bit, so if you believe this may be specific for my case, maybe this change isn't necessary or should be it's own section.